Client userid addded

+++ Back/src/services/client.service.js
@@ -2,6 +2,7 @@
 const jwt = require('jsonwebtoken');
 const bcrypt = require('bcryptjs');
 const db = require('../_helpers/db');
+const { param } = require('../controllers/clients.controller');
 
 module.exports = {
     getAll,
@@ -11,27 +12,51 @@ module.exports = {
     delete: _delete
 };
 
-async function getAll() {
-    return await db.Client.findAll();
+async function getAll(user) {
+    return await db.Client.findAll({
+        where: {
+          userId: user.id
+        }
+      });
 }
 
-async function getById(id) {
-    return await getClient(id);
+async function getById(user, id) {
+    client = await db.Client.findOne({
+        where: {
+          id: id,
+          userId: user.id
+        },
+      });
+    if (!client) throw 'Client not found';
+    return await client;
 }
 
-async function create(params) {
+async function create(user, params) {
+    params.userId = user.id;
     await db.Client.create(params);
 }
 
-async function update(id, params) {
-    const client = await getClient(id);
+async function update(user, id, params) {
+    client = await db.Client.findOne({
+        where: {
+          id: id,
+          userId: user.id
+        },
+      });
+    if (!client) throw 'Client not found';
     Object.assign(client, params);
     await client.save();
     return client.get();
 }
 
-async function _delete(id) {
-    const client = await getClient(id);
+async function _delete(user, id) {
+    client = await db.Client.findOne({
+        where: {
+          id: id,
+          userId: user.id
+        },
+      });
+    if (!client) throw 'Client not found';
     await client.destroy();
 }
